Moved logic to mod package
