+++ cmd/sysl/cmd_mod.go
@@ -3,13 +3,11 @@ package main
 import (
 	"errors"
 	"fmt"
-	"os/exec"
 	"strings"
 
+	"github.com/anz-bank/sysl/pkg/mod"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
-
-const syslModuleName = "syslmodules"
 
 type modCmd struct {
 	subcommand string
@@ -27,21 +25,9 @@ func (m *modCmd) Configure(app *kingpin.Application) *kingpin.CmdClause {
 }
 
 func (m *modCmd) Execute(args ExecuteArgs) error {
-	//nolint:singleCaseSwitch
-	switch m.subcommand {
-	case "init":
-		return syslModInit(args)
+	if m.subcommand == "init" {
+		return mod.SyslModInit(args.Logger)
 	}
+
 	return errors.New("command not recognized")
-}
-
-func syslModInit(args ExecuteArgs) error {
-	// makes the assumption that the CWD is not a go module since we hijack this command
-	out, err := exec.Command("go", "mod", "init", syslModuleName).CombinedOutput()
-	if err != nil {
-		return err
-	}
-
-	args.Logger.Info(string(out))
-	return nil
 }
